+++ git_machete/cli.py
@@ -49,7 +49,7 @@ command_groups: List[Tuple[str, List[str]]] = [
 commands_and_aliases = list(long_docs.keys()) + list(alias_by_command.keys())
 
 
-def get_help_description(command: str = None) -> str:
+def get_help_description(display_help_topics: bool, command: str = None) -> str:
     usage_str = ''
     if command and command in commands_and_aliases:
         if command in long_docs:
@@ -65,6 +65,9 @@ def get_help_description(command: str = None) -> str:
             "    Get familiar with the help for <b>format</b>, <b>edit</b>,"
             " <b>status</b> and <b>update</b>, in this order.\n\n")
         for hdr, cmds in command_groups:
+            if not display_help_topics:
+                if hdr == 'General topics':
+                    cmds = [topic for topic in cmds if topic not in ['config', 'format', 'hooks']]
             usage_str += underline(hdr) + '\n\n'
             for cm in cmds:
                 alias = f", {alias_by_command[cm]}" if cm in alias_by_command else ""
@@ -114,7 +117,7 @@ class MacheteHelpAction(argparse.Action):
     ) -> None:
         # parser name (prog) is expected to be `git machete` or `git machete <command>`
         command_name = parser.prog.replace('git machete', '').strip()
-        print(get_help_description(command_name))
+        print(get_help_description(display_help_topics=True, command=command_name))
         parser.exit(status=ExitCode.SUCCESS)
 
 
@@ -518,13 +521,13 @@ def launch(orig_args: List[str]) -> None:
         set_utils_global_variables(cli_opts)
 
         if not orig_args:
-            print(get_help_description())
+            print(get_help_description(display_help_topics=False))
             exit_script(ExitCode.ARGUMENT_ERROR)
 
         cmd = parsed_cli.command
 
         if cmd == "help":
-            print(get_help_description(parsed_cli.topic_or_cmd))
+            print(get_help_description(display_help_topics=True, command=parsed_cli.topic_or_cmd))
             exit_script(ExitCode.SUCCESS)
         elif cmd == "version":
             version()